HUGE SHIT DONE

+++ Platformer/LevelManager.cs
@@ -1,6 +1,5 @@
 public class LevelManager
 {
-    public static List<Level> allLevels = new List<Level>();
     public static int currentLevel;
     public static void LoadLevel(int level)
     {
@@ -27,6 +26,22 @@ public class LevelManager
 
         }
     }
+    public static void LoadLevelToEdit(int level)
+    {
+        ClearLevel();
+        Level lvl = GetLevelJson(level);
+        // Console.WriteLine(lvl.level);
+        currentLevel = lvl.level;
+
+        // Reset player position to start position
+        new Tile(new Vector2(lvl.startPos.x, lvl.startPos.y), "player");
+
+        // Load tiles and enemies to gameobjects
+        foreach (JsonGameobject gameobject in lvl.gameobjects)
+        {
+            new Tile(new Vector2(gameobject.x, gameobject.y), gameobject.type);
+        }
+    }
     public static void ResetLevel()
     {
         LoadLevel(currentLevel);
@@ -36,6 +51,8 @@ public class LevelManager
         GameObject.gameObjects.RemoveAll(x => (x is Tile));
         GameObject.gameObjects.RemoveAll(x => (x is Enemy));
         GameObject.gameObjects.RemoveAll(x => (x is Collectible));
+
+        currentLevel = 0;
     }
     public static void SaveLevel()
     {
@@ -55,17 +72,32 @@ public class LevelManager
         startPos.x = (int)(player.rect.x / 16 / GameObject.scale);
         startPos.y = (int)(player.rect.y / 16 / GameObject.scale); // MAKE IT CHANGE!!!!
 
+        if (currentLevel == 0) currentLevel = FindFirstAvailableName();
+
         Level newLevel = new Level();
         newLevel.gameobjects = jsonObjects;
-        newLevel.level = (totalLevels + 1); // Set level name
+        newLevel.level = currentLevel; // Set level name
         newLevel.startPos = startPos;
 
         string jsonString = JsonSerializer.Serialize<Level>(newLevel);
 
-        string filePath = @"levels\" + (totalLevels + 1 + ".json");
-        File.Create(filePath).Dispose();
+        string filePath = @"levels\" + currentLevel + ".json";
+        if (!File.Exists(filePath))
+            File.Create(filePath).Dispose();
 
         File.WriteAllText(filePath, jsonString);
+    }
+    static int FindFirstAvailableName()
+    {
+        string[] files = Directory.GetFiles(@"levels\");
+
+        int availableInt = 1;
+        for (int i = 0; i < files.Length; i++)
+        {
+            if (availableInt != int.Parse(Path.GetFileNameWithoutExtension(files[i]))) return availableInt;
+            availableInt++;
+        }
+        return files.Length + 1;
     }
     static Level GetLevelJson(int lvl)
     {
